server/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Authlib==1.2.1
backoff==2.2.1
charset-normalizer==3.3.2
dataclasses-json==0.6.3
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-jazzmin==2.6.0
emoji==2.8.0
filetype==1.2.0
grpcio==1.60.0rc1
grpcio-tools==1.60.0rc1
langdetect==1.0.9
marshmallow==3.20.1
mypy-extensions==1.0.0
nltk==3.8.1
oauthlib==3.2.2
pdf2image==1.16.3
Pillow==10.1.0
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyPDF2==3.0.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python3-openid==3.2.0
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tabulate==0.9.0
typing-inspect==0.9.0
validators==0.22.0
weaviate-client==4.4b1
